Add infrastructure code to built singularity container
This commit includes code that containerize the library environment and run
notebooks in a simple way. A makefile is added to create one linear deployments.
To build the singularity container we rely on the `--remote` feature. Hence,
the built is depending on the users ability to create a singularity user.